% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hfdata}
\alias{hfdata}
\title{hfdata as sample data}
\format{\code{hfdata} consist of several columns:
\itemize{
  \item centre: The centre names and "Sabah" is the region name
  \item id: The identification number of the centres
  \item 2003-2007: The measurement collected yearly based
  \item case1: Cases normally distributed with mean=60 and SD=10
  \item case2: Cases normally distributed with mean=5 and SD=2
  \item extt: Variable with extreme values
}}
\usage{
hfdata
}
\description{
hfdata is just a randomly created dataset to show how this package works.
The centre names are derived from names of towns on the north-eastern part of Borneo.
}
\keyword{datasets}
