% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegData}
\alias{RegData}
\title{RegData as sample data}
\format{\code{RegData} consist of 2 columns:
\itemize{
  \item Variable: The specific aims or objectives
  \item Percentage: The percentage of completeness
}}
\usage{
RegData
}
\description{
A sample data for completeness of different aims in a projects. The data is based on
The Norwegian Childhood Diabetes Registry (NCDR). The variables are
as follows:
}
\keyword{datasets}
