% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reghist.R
\name{reghist}
\alias{reghist}
\title{Histogram with comparison to national data}
\usage{
reghist(data, x, y)
}
\arguments{
\item{data}{Data set}

\item{x}{x-axis}

\item{y}{y-axis}

\item{show}{The category that will have distinct colour f.eg. National mean}

\item{cut}{Where to split to show text inside or outside the bar eg. all under 10
will show text outside bar while all above 10 will show text inside bar}

\item{sort}{Sort data ascending order}

\item{title}{Title for the plot}

\item{ylab}{Label for y-axis}

\item{xlab}{Label for x-axis}
}
\description{
Create a histogram for each health institutions or health regions
compared to the national data.
}
