% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regline.R
\name{regline}
\alias{regline}
\title{Line plot to show trend}
\usage{
regline(data, x, y, grp, title, ylab, colp, digit, ...)
}
\arguments{
\item{data}{Data set}

\item{x}{x-axis}

\item{y}{y-axis}

\item{grp}{Group variable}

\item{title}{Title for the plot}

\item{ylab}{Label for y-axis}

\item{colp}{Color palettes to use from ColorBrewer. To check other palettes run
library(RColorBrewer); display.brewer.all()}

\item{digit}{Number of digit to show}

\item{...}{Additional arguments}
}
\description{
Create a line plot that can be used to elucidate if trends exit over time.
}
\examples{
regline(data = yrdata, x=year, y=pros, grp=var)
regline(yrdata, year, pros, var, colp="Set1", digit=1)

}
