% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcom.R
\name{regtab}
\alias{regtab}
\title{Table for plot}
\usage{
regtab(data, x, ...)
}
\arguments{
\item{data}{Data set}

\item{x}{x-axis}

\item{...}{Additional arguments}
}
\description{
Create table with column for dummy reference for plot with table
}
