\name{rreg-package}
\alias{rreg-package}
\alias{rreg}
\docType{package}
\title{
\packageTitle{rreg}
}
\description{
\packageDescription{rreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rreg}
\packageIndices{rreg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rreg}

Maintainer: \packageMaintainer{rreg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
