% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreg.R
\docType{package}
\name{rreg}
\alias{rreg}
\alias{rreg-package}
\title{\code{rreg} package}
\description{
Data visualization for Norwegian Health Quality Registries with R. This package
will assist and standardize the visualization of data from the Norwegian Health
Quality Registries. The standardization is based on the requirement specified by
the Nasjonalt servicemiljø for medisinske kvalitetsregistre.
}
\author{
Yusman Kamaleri <ybkamaleri@gmail.com>
}
