% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yrdata}
\alias{yrdata}
\title{yrdata is sample data for trends}
\format{\code{yrdata} consist of these variables:
\itemize{
  \item year: List of different years
  \item var: Variable to be grouped
  \item N: Number of n for each group
  \item sum: Total for each year
  \item pros: Percentage for each group
}}
\usage{
yrdata
}
\description{
yrdata is just a sample data to use in example for "regline" function.
}
\keyword{datasets}
