% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regbar.R
\name{regbar}
\alias{regbar}
\title{Barplot with comparison to national data}
\usage{
regbar(data, x, y, diff, ascending = TRUE, title, ylab, col1, col2,
  flip = TRUE, ...)
}
\arguments{
\item{data}{Data set}

\item{x}{x-axis}

\item{y}{y-axis}

\item{diff}{Differentiate a specific bar from the rest for a clear comparison
eg. National compared to the different districts}

\item{ascending}{Sort data ascending order}

\item{title}{Title for the plot}

\item{ylab}{Label for y-axis}

\item{col1}{Color for bars}

\item{col2}{Color for the 'diff' bar}

\item{flip}{Flip plot horizontally}

\item{...}{Additional arguments}
}
\description{
Create a barplot with the posibility to differentiate a specific item compared to
the rest. This is useful in a situation when there is a need to show the total
value as compared to each items in the x-axis. A specific example related to the
Norwegian Health Registries is when the sum of each health institutions or health
regions compared to the national data.
}
