% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcom.R
\name{regcom}
\alias{regcom}
\title{Barplot with point and table for comparison}
\usage{
regcom(data, x, yl, yc, tab = TRUE, ...)
}
\arguments{
\item{data}{Data set}

\item{x}{x-axis}

\item{yl}{Variable or column for local values}

\item{yc}{Variable or column for national values}

\item{tab}{Include table}

\item{...}{Additional arguments}
}
\description{
Create a barplot with point to visualise comparison. It is also possible to
include table to show the value of the plot.
}
